Changed approach. Moved version tracking to the class-identity-crisis.php
