+++ projects/packages/identity-crisis/composer.json
@@ -60,7 +60,7 @@
 		"mirror-repo": "Automattic/jetpack-identity-crisis",
 		"textdomain": "jetpack-idc",
 		"version-constants": {
-			"::PACKAGE_VERSION": "src/class-package-version.php"
+			"::PACKAGE_VERSION": "src/class-identity-crisis.php"
 		},
 		"changelogger": {
 			"link-template": "https://github.com/Automattic/jetpack-identity-crisis/compare/v${old}...v${new}"
