+++ projects/plugins/starter-plugin/changelog/add-idc-package-version-tracking#2
@@ -1,0 +1,5 @@
+Significance: patch
+Type: changed
+Comment: Updated composer.lock.
+
+
